+++ pkg/analysis_server/lib/src/services/completion/dart/local_reference_contributor.dart
@@ -609,7 +609,7 @@ class _LocalVisitor extends LocalDeclarationVisitor {
         ? NullabilitySuffix.none
         : NullabilitySuffix.star;
 
-    return element.instantiate2(
+    return element.instantiate(
       typeArguments: typeArguments,
       nullabilitySuffix: nullabilitySuffix,
     );